+++ driver-core/src/main/java/com/datastax/driver/core/querybuilder/Clause.java
@@ -81,12 +81,9 @@ public abstract class Clause extends Utils.Appendeable {
             // since in that case it's more elegant to use an equal. On the other side,
             // it is a lot more useful to do:
             //    ... IN ? ...
-            // which binds the variable to the full list the IN is on. Now, at the time
-            // of this writting, that last syntax is not supported, but it will be
-            // once CASSANDRA-4210 is fixed. So to avoid breaking compatibility later on,
-            // we support the most common behavior now.
-            if (values.size() == 1 && values.get(0) == QueryBuilder.BIND_MARKER) {
-                Utils.appendName(name, sb).append("IN ?");
+            // which binds the variable to the full list the IN is on.
+            if (values.size() == 1 && values.get(0) instanceof BindMarker) {
+                Utils.appendName(name, sb).append("IN ").append(values.get(0));
                 return;
             }