+++ driver-core/src/main/java/com/datastax/driver/core/querybuilder/Using.java
@@ -18,23 +18,49 @@ package com.datastax.driver.core.querybuilder;
 import java.nio.ByteBuffer;
 import java.util.List;
 
-public class Using extends Utils.Appendeable {
+public abstract class Using extends Utils.Appendeable {
 
-    private final String optionName;
-    private final long value;
+    final String optionName;
 
-    Using(String optionName, long value) {
+    private Using(String optionName) {
         this.optionName = optionName;
-        this.value = value;
     }
 
-    @Override
-    void appendTo(StringBuilder sb, List<ByteBuffer> variables) {
-        sb.append(optionName).append(" ").append(value);
+    static class WithValue extends Using {
+        private final long value;
+
+        WithValue(String optionName, long value) {
+            super(optionName);
+            this.value = value;
+        }
+
+        @Override
+        void appendTo(StringBuilder sb, List<ByteBuffer> variables) {
+            sb.append(optionName).append(" ").append(value);
+        }
+
+        @Override
+        boolean containsBindMarker() {
+            return false;
+        }
     }
 
-    @Override
-    boolean containsBindMarker() {
-        return false;
+    static class WithMarker extends Using {
+        private final BindMarker marker;
+
+        WithMarker(String optionName, BindMarker marker) {
+            super(optionName);
+            this.marker = marker;
+        }
+
+        @Override
+        void appendTo(StringBuilder sb, List<ByteBuffer> variables) {
+            sb.append(optionName).append(" ").append(marker);
+        }
+
+        @Override
+        boolean containsBindMarker() {
+            return true;
+        }
     }
 }